make TelegramBot.js as an library for GAS
