+++ Gulpfile.js
@@ -1,0 +1,11 @@
+const { src, dest } = require("gulp")
+const expose = require("gulp-expose")
+const rename = require("gulp-rename")
+const IDENTIFIER = "TelegramBot"
+
+exports.build = function () {
+    return src("src/*.js")
+        .pipe(expose("this", IDENTIFIER))
+        .pipe(rename({ extname: '.gs' }))
+        .pipe(dest("dist"))
+}
