--- appsscript.json
@@ -1,6 +1,0 @@
-{
-  "timeZone": "Asia/Kolkata",
-  "exceptionLogging": "STACKDRIVER",
-  "runtimeVersion": "V8",
-  "dependencies": {}
-}
